[0.9.0-SNAPSHOT] StringExtensions.kt impl added BtcBlock.kt and Block.kt models fixed BlockHashCalculator.kt contract fixed for empty or string cause Mono<String> Tests fixed and updated

+++ src/main/kotlin/com/arrival/crypto/bitcoinanalyzer/model/Block.kt
@@ -11,7 +11,7 @@ import java.math.BigInteger
 open class Block(
     open val height: Long,
     open val size: Int,
-    open val timestamp: Int,
+    open val timestamp: Long,
     open val hash: String
 ) : Comparable<Block> {
